Offset should be Long

+++ bot-api/src/main/java/org/yanex/telegram/TelegramBotService.kt
@@ -175,7 +175,7 @@ interface TelegramBotService {
 
     @GET("/getUpdates")
     fun getUpdates(
-            @Query("offset") offset: Int? = null,
+            @Query("offset") offset: Long? = null,
             @Query("limit") limit: Int? = null,
             @Query("timeout") timeout: Int? = null
     ): Response<List<Update>>
